Add an About page to Settings and the docs
The About section of Settings moves into its own component and gains two
short blocks in the dialog's plain style: who builds Muxus, with links to
flosch.me, GitHub and LinkedIn, and how to support it, with a Buy Me a
Coffee button and a link to star the repository. The build line now names
the platform and the licence, and the app's own links (documentation,
source, releases) sit under it. The update controls are unchanged.

The docs get an About tab with the same content in prose: who makes Muxus,
the personal links, the support buttons, the other projects and how to get
in touch. The landing page and the Community page use the same icon
buttons and link through to it, and the footer carries the Buy Me a Coffee
and website icons next to GitHub.
