+++ docs/about.md
@@ -1,0 +1,50 @@
+---
+icon: lucide/user-round
+hide:
+  - navigation
+---
+
+# About
+
+Muxus is built by me (FloSch), in the open and in my spare time. I build open-source tooling
+for network automation and the desktop apps that go with it, and Muxus is the terminal client I
+wanted for that work: the hosts from `~/.ssh/config`, split panes, SFTP, tunnels and a remote
+editor on one connection, in a window I can save and reopen.
+
+- :material-web: [flosch.me](https://flosch.me/), who I am and what I work on
+- :simple-github: [FloSch62 on GitHub](https://github.com/FloSch62), everything I build, Muxus included
+- :fontawesome-brands-linkedin: [LinkedIn](https://www.linkedin.com/in/florian-schwarz-812a34145/), say hello
+
+## Support Muxus
+
+Muxus is free and stays free. If it saves you time, a coffee keeps the releases coming, and a
+star helps others find it.
+
+[:simple-buymeacoffee: Buy me a coffee](https://www.buymeacoffee.com/FloSch62){ .md-button .md-button--primary }
+[:material-star-outline: Star on GitHub](https://github.com/FloSch62/muxus){ .md-button }
+
+## Other things I build
+
+<div class="grid cards" markdown>
+
+-   :material-kubernetes: **[Kubus](https://github.com/FloSch62/Kubus)**
+
+    ---
+
+    A free, open-source Kubernetes GUI. Every cluster and resource, logs, shells, port forwards,
+    metrics and Helm in one local app.
+
+-   :material-lan: **[containerlab tooling](https://github.com/srl-labs/vscode-containerlab)**
+
+    ---
+
+    The VS Code extension and the desktop app for containerlab, which spins up container-based
+    networking labs from topology files.
+
+</div>
+
+## Get in touch
+
+Bug reports, ideas and pull requests are always welcome on the
+[issue tracker](https://github.com/FloSch62/muxus/issues). For anything else, the links above
+reach me.
