+++ zensical.toml
@@ -72,6 +72,7 @@ nav = [
     { "Building from source" = "community/development.md" },
     { "FAQ" = "community/faq.md" },
   ] },
+  { "About" = "about.md" },
 ]
 
 # ----------------------------------------------------------------------------
@@ -129,6 +130,16 @@ code = "JetBrains Mono"
 [[project.extra.social]]
 icon = "fontawesome/brands/github"
 link = "https://github.com/FloSch62/muxus"
+
+[[project.extra.social]]
+icon = "simple/buymeacoffee"
+link = "https://www.buymeacoffee.com/FloSch62"
+name = "Buy me a coffee"
+
+[[project.extra.social]]
+icon = "material/web"
+link = "https://flosch.me/"
+name = "flosch.me"
 
 # ----------------------------------------------------------------------------
 # Markdown extensions use the same stack as Material-based sites.
